move: AuthUser 관련 클래스 위치 변경 및 패키지 정리
- AuthUser와 관련된 클래스들을 적절한 패키지(dto.annotation)로 이동하여 패키지 구조 정리.
- 관련된 import 경로 수정.
