+++ app-api/src/main/java/org/yechan/api/ShowController.java
@@ -1,14 +1,14 @@
 package org.yechan.api;
 
-import static org.yechan.api.port.AuthUser.UserType.SELLER;
+import static org.yechan.dto.annotation.AuthUser.UserType.SELLER;
 
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.yechan.api.port.AuthUser;
 import org.yechan.api.port.ShowRegisterUseCase;
+import org.yechan.dto.annotation.AuthUser;
 import org.yechan.dto.request.ShowRegisterRequest;
 import org.yechan.dto.response.ShowRegisterResponse;
 import org.yechan.entity.Seller;
